% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWRI.palettes.R
\name{scale_fill_SWRI}
\alias{scale_fill_SWRI}
\title{scale_fill_SWRI function}
\usage{
scale_fill_SWRI(palette, discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{name, discrete true or false, reverse true or false}
}
\value{
colour palette
}
\description{
function to create colour gradients for "fill" in ggplot
}
