% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWRI.palettes.R
\name{SWRI_colours}
\alias{SWRI_colours}
\title{SWRI_colours function}
\usage{
SWRI_colours(...)
}
\arguments{
\item{The}{name of the colour(s)}
}
\value{
the hexadecimal code of the colour
}
\description{
returns hex code from colour descriptor
}
