% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWRI.palettes.R
\name{SWRI_pal}
\alias{SWRI_pal}
\title{SWRI_pal function}
\usage{
SWRI_pal(palette, reverse = FALSE, ...)
}
\arguments{
\item{palette}{name, reverse true or false}
}
\value{
colour palette
}
\description{
function to palettise a list of colours in palette, including gradients and reverse
}
